image-generator/requirements.txt
gradio
requests
Pillow
python-dotenv